+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 # Next Release
 
+- [#723](https://github.com/IAMconsortium/pyam/pull/723) Ensure correct order of `time` attribute
+
 # Release v1.7.0
 
 ## Highlights